#33 Change directory of Inventory file
