+++ bike-erp/src/components/Inventory/inventoryStyles.ts
@@ -1,27 +1,28 @@
 import { makeStyles } from "@material-ui/core";
 
 const useStyles = makeStyles((theme) => ({
-    title :{
+    title: {
         textAlign: "center",
         fontWeight: "bold",
-        fontFamily: "Inter"
+        fontFamily: "Inter",
+        fontSize: "64"
     },
-   background: {
-       background: "#DDDDDD"
-   },
-   tableHead: {
-       background: "#BFBFBF",
-   },
-   tableStyle:{
+    background: {
+        background: "#DDDDDD"
+    },
+    tableHead: {
+        background: "#BFBFBF",
+    },
+    tableStyle: {
         width: "90%",
         marginLeft: "auto",
         marginRight: "auto",
         border: "1px solid black"
-   },
-   innerTable:{
-       background: "#FFFFFF"
-   },
-    topRow:{
+    },
+    innerTable: {
+        background: "#FFFFFF"
+    },
+    topRow: {
         fontWeight: "bold",
         fontFamily: "Inter"
     }
