+++ bike-erp/src/test/Inventory/Inventory.test.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Inventory from '../../components/inventory';
+import Inventory from '../../components/Inventory/inventory';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 //import renderer from 'react-test-renderer';
